+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
     *Manuel Puyol*
 
+* **Breaking change**: Update `FlashComponent` `variant` argument to be `scheme`.
+
+    *Manuel Puyol*
+
 ## 0.0.33
 
 * Remove `TabbedComponent` validation requiring a tab to be selected.